switch to mysql

+++ db/migrate/20220512093402_add_order_to_line_item.rb
@@ -1,6 +1,6 @@
 class AddOrderToLineItem < ActiveRecord::Migration[6.1]
   def change
     add_reference :line_items, :order, null: true, foreign_key: true
-    change_column :line_items, :cart_id, :integer, null: true
+    change_column :line_items, :cart_id, :bigint, null: true
   end
 end
